+++ feature/utilcode/pkg/src/main/java/com/blankj/utilcode/pkg/feature/CoreUtilActivity.kt
@@ -44,6 +44,7 @@ import com.blankj.utilcode.pkg.feature.snackbar.SnackbarActivity
 import com.blankj.utilcode.pkg.feature.spStatic.SPStaticActivity
 import com.blankj.utilcode.pkg.feature.span.SpanActivity
 import com.blankj.utilcode.pkg.feature.toast.ToastActivity
+import com.blankj.utilcode.pkg.feature.uiMessage.UiMessageActivity
 import com.blankj.utilcode.pkg.feature.vibrate.VibrateActivity
 import com.blankj.utilcode.util.CollectionUtils
 import com.blankj.utilcode.util.UtilsTransActivity
@@ -190,6 +191,9 @@ class CoreUtilActivity : CommonActivity() {
                             activity.findViewById<TextView>(R.id.utilActionLoadingMsgTv).text = "Trans Activity is showing..."
                         }
                     })
+                },
+                CommonItemClick(R.string.demo_uiMessage, true) {
+                    UiMessageActivity.start(this)
                 },
                 CommonItemClick(R.string.demo_vibrate, true) {
                     VibrateActivity.start(this)