+++ lib/utilcode/src/main/java/com/blankj/utilcode/util/ConvertUtils.java
@@ -38,12 +38,34 @@ import java.util.List;
  */
 public final class ConvertUtils {
 
-    private static final int    BUFFER_SIZE = 8192;
-    private static final char[] hexDigits   =
+    private static final int    BUFFER_SIZE      = 8192;
+    private static final char[] HEX_DIGITS_UPPER =
             {'0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A', 'B', 'C', 'D', 'E', 'F'};
+    private static final char[] HEX_DIGITS_LOWER =
+            {'0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f'};
 
     private ConvertUtils() {
         throw new UnsupportedOperationException("u can't instantiate me...");
+    }
+
+    /**
+     * Int to hex string.
+     *
+     * @param num The int number.
+     * @return the hex string
+     */
+    public static String int2HexString(int num) {
+        return Integer.toHexString(num);
+    }
+
+    /**
+     * Hex string to int.
+     *
+     * @param hexString The hex string.
+     * @return the int
+     */
+    public static int hexString2Int(String hexString) {
+        return Integer.parseInt(hexString, 16);
     }
 
     /**
@@ -130,7 +152,20 @@ public final class ConvertUtils {
      * @return hex string
      */
     public static String bytes2HexString(final byte[] bytes) {
+        return bytes2HexString(bytes, true);
+    }
+
+    /**
+     * Bytes to hex string.
+     * <p>e.g. bytes2HexString(new byte[] { 0, (byte) 0xa8 }, true) returns "00A8"</p>
+     *
+     * @param bytes       The bytes.
+     * @param isUpperCase True to use upper case, false otherwise.
+     * @return hex string
+     */
+    public static String bytes2HexString(final byte[] bytes, boolean isUpperCase) {
         if (bytes == null) return "";
+        char[] hexDigits = isUpperCase ? HEX_DIGITS_UPPER : HEX_DIGITS_LOWER;
         int len = bytes.length;
         if (len <= 0) return "";
         char[] ret = new char[len << 1];